+++ payment_lyra/controllers/__init__.py
@@ -7,4 +7,4 @@
 # Copyright: Copyright © Lyra Network
 # License:   http://www.gnu.org/licenses/agpl.html GNU Affero General Public License (AGPL v3)
 
-import main
+from . import main